models/requirements.txt
ultralytics
opencv-python
numpy
onnxruntime
torch
torchvision
pyyaml
fastapi
uvicorn[standard]
redis
boto3
websockets
